(PDK-1090) Test task names (#18)
Prior to this commit the task names were not validated. This commit
updates the creates a rake task to check that the puppet task names
are compliant with the naming rules. *Note* that this is a workaround to
PDK-1090 and does not affect `pdk validate`. To use it run `bundle exec
rake tasknames`.