(maint) Add KB Number tasks display on the forge
Prior to this commit, the task view on the forge did not differentiate
between the tasks on the page. This made it very difficult to find out
the details for a particular task. This commit adds the task name to the
beginning of the description. This commit also adds the tasks version
number for the metadata files that were missing it.