(SUP-2054) Update module to support PE 2019.x. (#38)
* (maint) Enable support for Puppet 6.x

Prior to this commit, the module did not support puppet 6.x. This commit
changes the metadata to add support for any puppet below 7. This enables
support for PE 2019.x

* (maint) Update module to use PDK 1.10.0

Prior to this commit, the module was using an older version of PDK. This
commit updates the module to use PDK 1.10.0.